Magic happens here
Removed most of the BinaryReaders and bitstreams.
IBitStream provides everything we need.

Yes, the bitstreams for packet entity parsing still exist. Just be patient.
Right now, the new demo packet parsing is strictly experimental and just for
a single packet (CNETMsg_Tick). To provide a nice transition, all unhandled
messages will be forwarded to the old protobuf-net code. Additionally, this
allows us to be really lazy and just implement te most frequent packets.

We also need to replace IMessageParser, as the parsers have to accept things
that aren't IExtensible and preferrably not as objects, to avoid the boxing
allocation. That's also why I chose the tick message: Because nobody uses it.

I think I'll just call the parsers directly for now until someone comes up
with a nice *and fast* solution. /cc @moritzuehling
Yes, this breaks the nice abstraction of extensible message parsing, but did
anyone even use that?